pyproject.toml
[project]
name = "human-captcha"
version = "1.0.0"
description = "Motion-based differential perception CAPTCHA demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "numpy>=1.24.0",
    "numba>=0.58.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["captcha"]
